requirements.txt
absl-py==2.4.0
asciitree==0.3.3
astunparse==1.6.3
cellpose==3.1.1.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
csbdeep==0.8.2
cycler==0.12.1
Deprecated==1.3.1
donfig==0.8.1.post1
fasteners==0.20
fastremap==1.18.0
filelock==3.25.2
fill_voids==2.1.1
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
gast==0.7.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
google-crc32c==1.8.0
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.16.0
idna==3.11
imagecodecs==2026.3.6
ImageIO==2.37.3
Jinja2==3.1.6
keras==2.15.0
kiwisolver==1.5.0
lazy-loader==0.5
libclang==18.1.1
llvmlite==0.46.0
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
ml-dtypes==0.2.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.6.1
numba==0.64.0
numcodecs==0.15.1
numpy==1.26.4
oauthlib==3.3.1
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
packaging==25.0
pillow==12.1.1
protobuf==4.25.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
PySide6==6.11.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.0
requests-oauthlib==2.0.0
roifile==2026.2.10
scikit-image==0.26.0
scipy==1.17.1
segment-anything==1.0
shiboken6==6.11.0
six==1.17.0
stardist==0.9.2
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
tifffile==2026.3.3
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
wrapt==1.14.2
zarr==2.18.7
einops==0.8.2
instanseg-torch==0.1.1
